% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.apv.R
\name{growthfd.apv}
\alias{growthfd.apv}
\title{Compute apv of model instance}
\usage{
growthfd.apv(model, par)
}
\arguments{
\item{model}{FPCA growth model}

\item{par}{Params of the model, corresponding to some individual}
}
\value{
Age of maximum growth velocity
}
\description{
This function computes apv related to the certain instance of the model
described by the given parameters.
}
