% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd.ApvRegVelocity}
\alias{growthfd.ApvRegVelocity}
\title{Register a velocity curve at population apv}
\usage{
growthfd.ApvRegVelocity(model, par)
}
\arguments{
\item{model}{FPCA growth model}

\item{par}{Params of the model, corresponding to the individual}
}
\value{
Velocity at apv data frame
}
\description{
This function registers a curve corresponding to the supplied parameters 
onto the population apv.
}
\examples{
filename <- system.file("extdata", "data.csv", package="growthfd", mustWork=TRUE)
csv <- read.csv(filename)
d <- data.frame('id'=as.factor(csv[,'id']), 'x'=csv[,'age'], 'y'=csv[,'height'])
m <- d$id == 'John'
fit <- growthfd.fit(model.bgs.m, age=d$x[m], height=d$y[m])
data<-growthfd.ApvRegVelocity(model.bgs.m, fit$par)
}
