% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.dropoutsIds.Height}
\alias{growthfd.bgs.dropoutsIds.Height}
\title{List ids of individuals to be dropped from height modeling}
\usage{
growthfd.bgs.dropoutsIds.Height()
}
\value{
Vector of ids
}
\description{
This function returns a vector containing ids of individuals with incomplete 
stature measurements.
}
