% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.evalMonotone}
\alias{growthfd.bgs.evalMonotone}
\title{Evaluate monotone fda splines}
\usage{
growthfd.bgs.evalMonotone(fda, age, deriv = 0)
}
\arguments{
\item{fda}{Fda object}

\item{age}{Vector of ages to be evaluated}

\item{deriv}{Derivative of the curve}
}
\value{
Matrix of evaluated points
}
\description{
This function evaluates the set of monotone splines.
}
