% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.gather}
\alias{growthfd.bgs.gather}
\title{Gather selected columns}
\usage{
growthfd.bgs.gather(data, prefix = "vysk", age = NULL)
}
\arguments{
\item{data}{BGS data}

\item{prefix}{Columns prefix}

\item{age}{Vector containing ages (optional)}
}
\value{
Gathered data
}
\description{
Selects columns with given prefix for supplied ages and gathers them into 
a matrix together with id and sex data
}
