% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.interpolateNAs}
\alias{growthfd.bgs.interpolateNAs}
\title{Estimate NA values}
\usage{
growthfd.bgs.interpolateNAs(gatheredData)
}
\arguments{
\item{gatheredData}{Data in gathered form}
}
\value{
Interpolated data
}
\description{
Interpolates missing values using spline method. Interpolates data 
from the 'value' column and join them as the 'valuei' column.
}
