% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.invertTw}
\alias{growthfd.bgs.invertTw}
\title{Compute inverse time-warping functions}
\usage{
growthfd.bgs.invertTw(age, tw)
}
\arguments{
\item{age}{Vector of ages}

\item{tw}{Fda object containing the time-warping functions}
}
\value{
Fda object containing the inverse functions
}
\description{
Computes inverse for given functions.
}
