% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.measurementsAge}
\alias{growthfd.bgs.measurementsAge}
\title{List ages of measurements}
\usage{
growthfd.bgs.measurementsAge()
}
\value{
Vector of ages
}
\description{
This function returns a vector of ages when the measurements were performed.
}
