% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.model}
\alias{growthfd.bgs.model}
\title{Create FPCA growth model}
\usage{
growthfd.bgs.model(amplitude, itw, nharm = 6)
}
\arguments{
\item{itw}{Fda object of inverse time warping functions}

\item{nharm}{Number of harmonic functions for each fpca}

\item{ampitude}{Fda object of registered growth functions}
}
\value{
FPCA growth model
}
\description{
Creates FPCA growth model from fda objects of growth functions registered
on apv and inverse time warping functions.
}
