% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.plotAll}
\alias{growthfd.bgs.plotAll}
\title{Plot curves in one figure}
\usage{
growthfd.bgs.plotAll(age, values, xlimit = NULL, ylimit = NULL)
}
\arguments{
\item{age}{Vector of ages}

\item{values}{Matrix containing curves as columns}

\item{xlimit}{Limits for the x axis}

\item{ylimit}{Limits for the y axis}
}
\value{
GGPlot2 plot
}
\description{
Plots all curves from given matrix into a single figure.
}
