% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.plotFda}
\alias{growthfd.bgs.plotFda}
\title{Plot all curves to pdf}
\usage{
growthfd.bgs.plotFda(
  age,
  ids,
  apvs,
  values,
  vel,
  acc,
  data,
  filename = "plots.pdf"
)
}
\arguments{
\item{age}{Vector of ages}

\item{ids}{Vector containing ids}

\item{apvs}{Vector containing apv for each individual}

\item{values}{Matrix with acceleration curves}

\item{vel}{Matrix with velocity curves}

\item{data}{Matrix with original data points}

\item{filename}{File name of the output pdf}
}
\description{
Plots value, velocity and acceleration curves together with apvs
and measured data points into separate figure for each individual. All plots
are stored into a single pdf file, one figure per page.
}
