% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.plotIndividualsPoints}
\alias{growthfd.bgs.plotIndividualsPoints}
\title{Plot individual points to pdf}
\usage{
growthfd.bgs.plotIndividualsPoints(ids, data, filename = "plots_points.pdf")
}
\arguments{
\item{ids}{Vector containing ids}

\item{data}{Matrix with original data points}

\item{filename}{File name of the output pdf}

\item{age}{Vector of ages}

\item{apvs}{Vector containing apv for each individual}

\item{vel}{Matrix with velocity curves}

\item{values}{Matrix with acceleration curves}
}
\description{
Plots measured data points into separate figure for each individual. All plots
are stored into a single pdf file, one figure per page.
}
