% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.registerCurvesToApvs}
\alias{growthfd.bgs.registerCurvesToApvs}
\title{Register curves to the apvs}
\usage{
growthfd.bgs.registerCurvesToApvs(fdaObject, apvs)
}
\arguments{
\item{fdaObject}{Fda object contaning the curves}

\item{apvs}{Vector containing the respective apvs}
}
\value{
FDA object containing the time warping functions
}
\description{
Calculates the time warping functions with respect to the supplied apvs 
and the growth curves for the final refinement.
}
