% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.resample}
\alias{growthfd.bgs.resample}
\title{Resample the data}
\usage{
growthfd.bgs.resample(interpolatedData)
}
\arguments{
\item{interpolatedData}{Data to be resampled.}
}
\value{
Resampled data
}
\description{
Resample the data without NA values to fine grid.
}
