% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd.evaluate}
\alias{growthfd.evaluate}
\title{Generate a Discrete Growth Curve}
\usage{
growthfd.evaluate(x, par, model, deriv = 0)
}
\arguments{
\item{x}{Ages to be evaluated}

\item{par}{Parameters of the model}

\item{model}{FPCA growth model}

\item{deriv}{Path to the input file}
}
\value{
Y-values of the evaluated curve
}
\description{
This function evaluates a curve function for given ages. Depending 
on a degree of derivation, the function produces stature, velocity 
or acceleration curve.
}
