% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd.fit}
\alias{growthfd.fit}
\title{Fit a FPCA Growth Curve Model to measurements of a single individual}
\usage{
growthfd.fit(model, age, height, nprint = 1)
}
\arguments{
\item{model}{FPCA growth model to be fitted}

\item{age}{Age at measured data points}

\item{height}{Height at at measured data points}

\item{nprint}{Verbosity}
}
\value{
An optimization result object
}
\description{
This function fits a model to the given measured data of a single individual.
}
\examples{
age <- c(6.9, 8.2, 10, 12.1)
height <- c(114, 122, 130, 141)
fit <- growthfd.fit(model.bgs.m, age=c(6.9, 8.2, 10, 12.1), height=c(114, 122, 130, 141))
x11()
growthfd.plot(model.bgs.m, fit$par)
points(age, height)
x11()
growthfd.plot(model.bgs.m, fit$par, from=0.5, deriv = 1)
x11()
growthfd.plot(model.bgs.m, fit$par, from=0.5, deriv = 2)

}
