% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.modelPars.R
\name{growthfd.modelPars}
\alias{growthfd.modelPars}
\title{Standardized model scores}
\usage{
growthfd.modelPars(model)
}
\arguments{
\item{model}{FPCA-based growth model}
}
\value{
Matrix containing the scores
}
\description{
This function returns model parameters for the individuals used for training
the model.
}
