% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd.plot.ApvRegVelocity}
\alias{growthfd.plot.ApvRegVelocity}
\title{Plot a velocity curve registered at apv}
\usage{
growthfd.plot.ApvRegVelocity(model, par)
}
\arguments{
\item{model}{FPCA growth model}

\item{par}{Params of the model, corresponding to the individual}
}
\value{
Velocity at apv plot
}
\description{
This function plots a velocity curve, registered at population (model) apv
in comparison with the mean curve.
}
\examples{
filename <- system.file("extdata", "data.csv", package="growthfd", mustWork=TRUE)
csv <- read.csv(filename)
d <- data.frame('id'=as.factor(csv[,'id']), 'x'=csv[,'age'], 'y'=csv[,'height'])
m <- d$id == 'John'
fit <- growthfd.fit(model.bgs.m, age=d$x[m], height=d$y[m])
p<-growthfd.plot.ApvRegVelocity(model.bgs.m, fit$par)
x11()
p
}
