% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.plot.R
\name{growthfd.plot}
\alias{growthfd.plot}
\title{Plot a Growth Curve}
\usage{
growthfd.plot(model, par, deriv = 0, from = 0, to = 18)
}
\arguments{
\item{model}{FPCA growth model}

\item{par}{Parameters of the model}

\item{deriv}{Path to the input file}

\item{from}{The lower age limit}

\item{to}{The upper age limit}
}
\description{
This function plots a stature, velocity or acceleration curve.
}
