% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.plotwarps}
\alias{growthfd.plotwarps}
\title{Plot eigenwarps and eigenamplitudes of the model}
\usage{
growthfd.plotwarps(model, deriv = 0, ylim = NULL)
}
\arguments{
\item{model}{FPCA growth model}

\item{deriv}{Derivation of the growth curve}

\item{ylim}{Limits the scale of y axis}
}
\description{
Plots the effects of individual parameters on the curve shape.
}
