% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd}
\alias{growthfd}
\title{Fit a FPCA Growth Curve Model to a population}
\usage{
growthfd(data, x, y, id, model, verbose = 1)
}
\arguments{
\item{data}{Data frame containing age, height and id of individuals}

\item{x}{Age at measured data points}

\item{y}{Height at measured data points}

\item{id}{Corresponding individual's id at measured data points}

\item{model}{FPCA growth model to be fitted}

\item{verbose}{Verbosity}
}
\value{
List containing individuals id and model scores
}
\description{
This function fits a model to the given measured data of a population.
}
