% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.RegVelocities.R
\name{growthfd.RegVelocities}
\alias{growthfd.RegVelocities}
\title{Prepare data for velocity boxplots registered on apv}
\usage{
growthfd.RegVelocities(model, par, ages, rndn = 0)
}
\arguments{
\item{model}{Model}

\item{par}{Parameters of the model fitted to the measurements}

\item{ages}{Ages of measurements points}

\item{rndn}{Count of random curves to be evaluated}
}
\value{
Data frames for population and the individual
}
\description{
This function prepares data for boxplots in time of measurements, 
after registration of the individual on population apv.
}
