% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd.residuals}
\alias{growthfd.residuals}
\title{Compute residuals}
\usage{
growthfd.residuals(x, y, par, model)
}
\arguments{
\item{x}{Vector with input ages}

\item{y}{Vector with target height measurements}

\item{par}{Parameters of the model}

\item{model}{FPCA growth model}
}
\value{
A vector of residuals
}
\description{
This function computes residuals between measured stature data 
and data generated from the growth model.
}
