% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.std.R
\name{growthfd.std}
\alias{growthfd.std}
\title{Generate a Curve Function}
\usage{
growthfd.std(par, model)
}
\arguments{
\item{par}{Phase and amplitude parameters}

\item{model}{FPCA growth model}
}
\value{
FDA function object
}
\description{
This function generates a growth curve function based on given model 
and parameters, describing the growth phase and amplitude.
}
