% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.warpfd.R
\name{growthfd.warpfd}
\alias{growthfd.warpfd}
\title{Time warping function}
\usage{
growthfd.warpfd(par, model)
}
\arguments{
\item{par}{Parameters of the model}

\item{model}{FPCA growth model}
}
\description{
This function returns the time warping function corresponding 
to supplied model and particular parameters.
}
