% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.warpfdInv.R
\name{growthfd.warpfdInv}
\alias{growthfd.warpfdInv}
\title{Inverse time warping function}
\usage{
growthfd.warpfdInv(par, model)
}
\arguments{
\item{par}{Parameters of the model}

\item{model}{FPCA growth model}
}
\description{
This function returns the *inverse* time warping function corresponding 
to supplied model and particular parameters.
}
