% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model.bgs.f}
\alias{model.bgs.f}
\title{FPCA model for girls}
\format{
An object of class \code{list} of length 3.
}
\usage{
model.bgs.f
}
\description{
Model trained using 167 female individuals from Brno Growth Study (BGS).
}
\keyword{datasets}
