% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model.bgs.m}
\alias{model.bgs.m}
\title{FPCA model for boys}
\format{
An object of class \code{list} of length 3.
}
\usage{
model.bgs.m
}
\description{
Model trained using 167 male individuals from Brno Growth Study (BGS).
}
\keyword{datasets}
