% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.bgs.R
\name{growthfd.bgs.smooth}
\alias{growthfd.bgs.smooth}
\title{Fit the monotone spline}
\usage{
growthfd.bgs.smooth(
  resampledData,
  monotone = T,
  norder = 6,
  Lfdobj = 3,
  lambda = 0.05
)
}
\arguments{
\item{resampledData}{Data to be interpolated by monotone fda splines}
}
\value{
Object with fitted splines
}
\description{
This function fit the monotone splines to the data.
}
