% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfd.R
\name{growthfd}
\alias{growthfd}
\title{Fit a FPCA Growth Curve Model to a population}
\usage{
growthfd(
  data,
  x,
  y,
  id,
  model,
  verbose = 1,
  bounds = "negative",
  filename = "",
  startFromId = NULL,
  parallel = F,
  scores.filename = "parallel.txt"
)
}
\arguments{
\item{data}{Data frame containing age, height and id of individuals}

\item{x}{Age at measured data points}

\item{y}{Height at measured data points}

\item{id}{Corresponding individual's id at measured data points}

\item{model}{FPCA growth model to be fitted}

\item{verbose}{Verbosity}

\item{bounds}{Limitation of the interval for milestones estimation, 'negative' or 'inverse'}

\item{filename}{File name for saving results after each individual}

\item{startFromId}{Start the evaluation from this id}

\item{parallel}{(Experimental) Parallel evaluation of the model fitting}

\item{scores.filename}{File name for continuous saving of the scores}
}
\value{
List containing individuals id and model
}
\description{
This function fits a model to the given measured data of a population.
}
\examples{
filename <- system.file("extdata", "data.csv", package="growthfd", mustWork=TRUE)
csv <- read.csv(filename)
d <- data.frame('id'=as.factor(csv[,'id']), 'x'=csv[,'age'], 'y'=csv[,'height'])
fit<-growthfd(data=d, x=x, y=y, id=id, model=model.bgs.m)

}
